sales_bot/requirements.txt
python-telegram-bot==20.0
openai==1.78.0
python-dotenv==1.1.0
aiohttp==3.11.18
sqlalchemy==2.0.40
aiosqlite==0.21.0
pytz==2025.2
python-dateutil==2.9.0
typing-extensions>=4.5.0
pydantic>=2.0.0